Revert "feat: add new for you feed (#4947)"
This reverts commit 4ccc0812ad7453f5c8feda9370d632da33f5721c.
