+++ apps/web/src/components/Home/FeedType.tsx
@@ -1,10 +1,7 @@
 import type { Dispatch, FC, SetStateAction } from 'react';
 
-import New from '@components/Shared/Badges/New';
-import isFeatureAvailable from '@helpers/isFeatureAvailable';
 import { Leafwatch } from '@helpers/leafwatch';
 import { HomeFeedType } from '@hey/data/enums';
-import { FeatureFlag } from '@hey/data/feature-flags';
 import { HOME } from '@hey/data/tracking';
 import { TabButton } from '@hey/ui';
 import { useProfileStore } from 'src/store/persisted/useProfileStore';
@@ -16,29 +13,23 @@ interface FeedTypeProps {
 
 const FeedType: FC<FeedTypeProps> = ({ feedType, setFeedType }) => {
   const { fallbackToCuratedFeed } = useProfileStore();
-  const enabled =
-    isFeatureAvailable(FeatureFlag.Gardener) ||
-    isFeatureAvailable(FeatureFlag.LensTeam);
 
   return (
     <div className="flex gap-3 overflow-x-auto px-5 sm:px-0">
-      {enabled && (
-        <TabButton
-          active={feedType === HomeFeedType.FORYOU}
-          badge={<New />}
-          name="For You"
-          onClick={() => {
-            setFeedType(HomeFeedType.FORYOU);
-            Leafwatch.track(HOME.SWITCH_FORYOU_FEED);
-          }}
-        />
-      )}
       <TabButton
         active={feedType === HomeFeedType.FOLLOWING}
         name={fallbackToCuratedFeed ? 'Curated Feed' : 'Following'}
         onClick={() => {
           setFeedType(HomeFeedType.FOLLOWING);
           Leafwatch.track(HOME.SWITCH_FOLLOWING_FEED);
+        }}
+      />
+      <TabButton
+        active={feedType === HomeFeedType.HIGHLIGHTS}
+        name="Highlights"
+        onClick={() => {
+          setFeedType(HomeFeedType.HIGHLIGHTS);
+          Leafwatch.track(HOME.SWITCH_HIGHLIGHTS_FEED);
         }}
       />
       <TabButton
