+++ apps/web/src/components/Home/index.tsx
@@ -2,30 +2,24 @@ import type { NextPage } from 'next';
 
 import NewPost from '@components/Composer/Post/New';
 import ExploreFeed from '@components/Explore/Feed';
-import isFeatureAvailable from '@helpers/isFeatureAvailable';
 import { Leafwatch } from '@helpers/leafwatch';
 import { HomeFeedType } from '@hey/data/enums';
-import { FeatureFlag } from '@hey/data/feature-flags';
 import { PAGEVIEW } from '@hey/data/tracking';
 import { GridItemEight, GridItemFour, GridLayout } from '@hey/ui';
 import { useEffect, useState } from 'react';
 import { useProfileStore } from 'src/store/persisted/useProfileStore';
 
 import FeedType from './FeedType';
-import ForYou from './ForYou';
 import Hero from './Hero';
+import Highlights from './Highlights';
 import PaidActions from './PaidActions';
 import Sidebar from './Sidebar';
 import Timeline from './Timeline';
 
 const Home: NextPage = () => {
-  const forYouEnabled =
-    isFeatureAvailable(FeatureFlag.Gardener) ||
-    isFeatureAvailable(FeatureFlag.LensTeam);
-
   const { currentProfile } = useProfileStore();
   const [feedType, setFeedType] = useState<HomeFeedType>(
-    forYouEnabled ? HomeFeedType.FORYOU : HomeFeedType.FOLLOWING
+    HomeFeedType.FOLLOWING
   );
 
   useEffect(() => {
@@ -45,8 +39,8 @@ const Home: NextPage = () => {
               <FeedType feedType={feedType} setFeedType={setFeedType} />
               {feedType === HomeFeedType.FOLLOWING ? (
                 <Timeline />
-              ) : feedType === HomeFeedType.FORYOU ? (
-                <ForYou />
+              ) : feedType === HomeFeedType.HIGHLIGHTS ? (
+                <Highlights />
               ) : feedType === HomeFeedType.PREMIUM ? (
                 <PaidActions />
               ) : null}
