+++ packages/data/tracking.ts
@@ -91,7 +91,7 @@ export const NOTIFICATION = {
 export const HOME = {
   SELECT_USER_FEED: 'Select user feed',
   SWITCH_FOLLOWING_FEED: 'Switch to following feed',
-  SWITCH_FORYOU_FEED: 'Switch to for you feed',
+  SWITCH_HIGHLIGHTS_FEED: 'Switch to highlights feed',
   SWITCH_PREMIUM_FEED: 'Switch to premium feed'
 };
