+++ packages/lens/apollo/cache/createFeedHighlightsFieldPolicy.ts
@@ -1,0 +1,9 @@
+import type { FieldPolicy } from '@apollo/client';
+
+import { cursorBasedPagination } from '../helpers';
+
+const createFeedHighlightsFieldPolicy = (): FieldPolicy => {
+  return cursorBasedPagination(['request', ['where']]);
+};
+
+export default createFeedHighlightsFieldPolicy;
