+++ packages/lens/documents/queries/FeedHighlights.graphql
@@ -1,0 +1,15 @@
+query FeedHighlights($request: FeedHighlightsRequest!) {
+  feedHighlights(request: $request) {
+    items {
+      ... on Post {
+        ...PostFields
+      }
+      ... on Quote {
+        ...QuoteFields
+      }
+    }
+    pageInfo {
+      next
+    }
+  }
+}
