--- packages/lens/documents/queries/ForYou.graphql
@@ -1,18 +1,0 @@
-query ForYou($request: PublicationForYouRequest!) {
-  forYou(request: $request) {
-    items {
-      publication {
-        ... on Post {
-          ...PostFields
-        }
-        ... on Quote {
-          ...QuoteFields
-        }
-      }
-      source
-    }
-    pageInfo {
-      next
-    }
-  }
-}
