+++ packages/ui/src/TabButton.tsx
@@ -6,7 +6,6 @@ import cn from '../cn';
 
 interface TabButtonProps {
   active: boolean;
-  badge?: ReactNode;
   className?: string;
   icon?: ReactNode;
   name: string;
@@ -17,7 +16,6 @@ interface TabButtonProps {
 
 const TabButton: FC<TabButtonProps> = ({
   active,
-  badge,
   className = '',
   icon,
   name,
@@ -50,7 +48,6 @@ const TabButton: FC<TabButtonProps> = ({
     >
       {icon}
       <span className={cn({ 'hidden sm:block': !showOnSm })}>{name}</span>
-      {badge}
     </button>
   );
 };
